+++ Editor/RoR2/AddressablesPathDictionary.cs
@@ -1,16 +1,16 @@
 using HG;
 using SimpleJSON;
 using System;
+using System.Collections;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Text.RegularExpressions;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
-using UnityEngine.ResourceManagement.AsyncOperations;
 using UnityEngine.ResourceManagement.ResourceLocations;
 using IOPath = System.IO.Path;
 
@@ -19,8 +19,159 @@ namespace RoR2.Editor
     /// <summary>
     /// The AddressablesPathDictionary is a struct that contains the metadata stored within the game's "lrapi_returns.json", which is a Dictionary of Addressable Path to GUID.
     /// </summary>
-    public struct AddressablesPathDictionary
+    public sealed class AddressablesPathDictionary
     {
+        /// <summary>
+        /// Represents an entry type within the dictionary
+        /// </summary>
+        public enum EntryType
+        {
+            /// <summary>
+            /// Represents a GUID entry.
+            /// </summary>
+            Guid,
+
+            /// <summary>
+            /// Represents a Path string entry
+            /// </summary>
+            Path
+        }
+
+        /// <summary>
+        /// Class utilized for looking up entries within the <see cref="AddressablesPathDictionary"/>.
+        /// <para></para>
+        /// This is an <see cref="IDisposable"/> class, be sure to call <see cref="Dispose"/> once you finish utilizing the lookup.
+        /// </summary>
+        public class EntryLookup : IDisposable
+        {
+            /// <summary>
+            /// The type of entries we're looking up
+            /// </summary>
+            public EntryType entryLookupType = EntryType.Path;
+
+            /// <summary>
+            /// Only entires that are of these types are returned, types are treated as OR.
+            /// </summary>
+            public Type[] typeRestriction = Array.Empty<Type>();
+
+            /// <summary>
+            /// An optional filter, only entires that contain this string are returned
+            /// </summary>
+            public string filter;
+
+            /// <summary>
+            /// The results of the lookup, this is populated after <see cref="PerformLookup"/> or <see cref="PerformLookupAsync"/>
+            /// </summary>
+            public ReadOnlyCollection<string> results;
+            public List<string> _results;
+
+            /// <summary>
+            /// Performs the Lookup with the given parameters
+            /// </summary>
+            /// <returns>The results of the lookup</returns>
+            public ReadOnlyCollection<string> PerformLookup()
+            {
+                IEnumerator sync = PerformLookupAsync();
+                while(sync.MoveNext())
+                {
+                }
+                return results;
+            }
+
+            /// <summary>
+            /// Performs the lookup with the given paramters in an async fashion
+            /// </summary>
+            /// <returns>A coroutine that can be awaited, once it's complete, the results are stored within <see cref="results"/></returns>
+            public IEnumerator PerformLookupAsync()
+            {
+                _results = ListPool<string>.RentCollection();
+
+                //Get each result for each type, request entries returns the cache if it exists, and if not, it builds it
+                foreach(Type type in typeRestriction)
+                {
+                    yield return null;
+                    string[] cache = AddressablesPathDictionary.GetInstance().RequestEntries(type, entryLookupType);
+
+
+                    var modulo = Mathf.Floor((Mathf.Log10(cache.Length) + 1) * 2);
+                    for (int i = 0; i < cache.Length; i++)
+                    {
+                        string cacheEntry = cache[i];
+
+                        if(i % modulo == 0)
+                        {
+                            yield return null;
+                        }
+
+                        //If we have a filter, filter the results
+                        if (!filter.IsNullOrEmptyOrWhiteSpace())
+                        {
+                            if(cacheEntry.Contains(filter) && !_results.Contains(cacheEntry))
+                            {
+                                _results.Add(cacheEntry);
+                            }
+                        }
+                        else
+                        {
+                            if(!_results.Contains(cacheEntry))
+                            {
+                                _results.Add(cacheEntry);
+                            }
+                        }
+                    }
+                }
+
+                //Store the results
+                results = new ReadOnlyCollection<string>(_results);
+                yield break;
+            }
+
+            /// <summary>
+            /// Sets <see cref="entryLookupType"/>'s value to <paramref name="entryType"/>
+            /// </summary>
+            /// <param name="entryType">The entry type to utilize for the lookup</param>
+            /// <returns>Itself</returns>
+            public EntryLookup WithLookupType(EntryType entryType)
+            {
+                entryLookupType = entryType;
+                return this;
+            }
+
+            /// <summary>
+            /// Sets <see cref="typeRestriction"/>'s values to the types passed in <paramref name="types"/>
+            /// </summary>
+            /// <param name="types">The Types to utilize as a restriction</param>
+            /// <returns>Itself</returns>
+            public EntryLookup WithTypeRestriction(params Type[] types)
+            {
+                typeRestriction = types;
+                return this;
+            }
+
+            /// <summary>
+            /// Sets <see cref="filter"/>'s values to the types passed in <paramref name="filter"/>
+            /// </summary>
+            /// <param name="filter">The filter to utilize</param>
+            /// <returns>Itself</returns>
+            public EntryLookup WithFilter(string filter)
+            {
+                this.filter = filter;
+                return this;
+            }
+
+            /// <summary>
+            /// Disposes the resources utilized for the lookup
+            /// </summary>
+            public void Dispose()
+            {
+                if(_results != null)
+                {
+                    ListPool<string>.ReturnCollection(_results);
+                    _results = null;
+                }
+            }
+        }
+
         private const string FILE_NAME = "lrapi_returns.json";
 
         private static string GetLrapiReturnsPath()
@@ -39,7 +190,7 @@ namespace RoR2.Editor
         /// <returns>The instance of AddressablesPathDictionary</returns>
         public static AddressablesPathDictionary GetInstance()
         {
-            if (!_instance.IsEmpty())
+            if (_instance != null && !_instance.IsEmpty())
             {
                 return _instance;
             }
@@ -81,6 +232,147 @@ namespace RoR2.Editor
         private string[] paths;
         private string[] guids;
 
+        #region Cache
+        private struct CacheHit
+        {
+            public string[] pathCache;
+            public string[] guidsCache;
+        }
+
+        private Dictionary<Type, CacheHit> _typeResultCache;
+
+        private string[] RequestEntries(Type[] types, EntryType entryType)
+        {
+            using var _0 = ListPool<string>.RentCollection(out var __result);
+            IEnumerable<string> result = __result;
+            using var _1 = ListPool<Type>.RentCollection(out var typesToProcess);
+
+            //First, we check for cached results
+            foreach(var type in types)
+            {
+                if(_typeResultCache.TryGetValue(type, out CacheHit cacheHit))
+                {
+                    switch(entryType)
+                    {
+                        case EntryType.Path: result.Union(cacheHit.pathCache); break;
+                        case EntryType.Guid: result.Union(cacheHit.guidsCache); break;
+                    }
+                    continue;
+                }
+                else
+                {
+                    //We've missed the cache for this type, so add it to the list of shit to check
+                    typesToProcess.Add(type);
+                }
+            }
+
+            //No types to process? return.
+            if(typesToProcess.Count == 0)
+            {
+                return result.ToArray();
+            }
+
+            foreach(var type in typesToProcess)
+            {
+                CacheHit cacheHit = BuildCacheForType(type);
+                switch(entryType)
+                {
+                    case EntryType.Path: result.Union(cacheHit.pathCache); break;
+                    case EntryType.Guid: result.Union(cacheHit.guidsCache); break;
+                }
+            }
+
+            return result.ToArray();
+        }
+        private string[] RequestEntries(Type type, EntryType entryType)
+        {
+            //First, we check if we have the cache
+            if(_typeResultCache.TryGetValue(type, out CacheHit cacheHit))
+            {
+                //We've hit the cache, return results
+                return entryType switch
+                {
+                    EntryType.Path => cacheHit.pathCache,
+                    EntryType.Guid => cacheHit.guidsCache,
+                    _ => Array.Empty<string>()
+                };
+            }
+
+            //We missed the cache, so time to build it for this type and return as needed
+            cacheHit = BuildCacheForType(type);
+            return entryType switch
+            {
+                EntryType.Path => cacheHit.pathCache,
+                EntryType.Guid => cacheHit.guidsCache,
+                _ => Array.Empty<string>()
+            };
+        }
+
+        private CacheHit BuildCacheForType(Type type)
+        {
+            using var _0 = ListPool<string>.RentCollection(out var resultPaths);
+            using var _1 = ListPool<string>.RentCollection(out var resultGuids);
+
+            //Work off GUIDS, like god intended
+            foreach(var guid in guids)
+            {
+                var resourceLocations = Addressables.LoadResourceLocationsAsync(guid).WaitForCompletion();
+
+                //There's a chance that there's no resources at this guid... idk???
+                if (resourceLocations.Count == 0)
+                {
+                    UnityEngine.Debug.Log($"No resources located at {GetPathFromGUID(guid)}");
+                    continue;
+                }
+
+                /* This is a bit fucky, but basically there's an issue where all sub-asset guids (the ones that have [] at the end) always has a pointer
+                *  back to its main asset, it's odd because even the non [] ones have it as well... specifically for sprites?
+                *  Anyways i need to figure some shit out
+                */
+
+                //If it contains the [, then stritctly match a sub-asset (at the very least the 2nd location).
+                bool strictSubAssetMatch = guid.Contains("[");
+                int resourceLocationIndex = -1;
+                if(strictSubAssetMatch && resourceLocations.Count > 1)
+                {
+                    resourceLocationIndex = 1;
+                }
+                else
+                {
+                    resourceLocationIndex = 0;
+                }
+
+                IResourceLocation resourceLocation = null;
+                try
+                {
+                    resourceLocation = resourceLocations[resourceLocationIndex];
+                }
+                catch(Exception e)
+                {
+                    UnityEngine.Debug.LogException(e);
+                    throw;
+                }
+                var resourceType = resourceLocation.ResourceType;
+
+                //If we obtained the resource type, and the resource type is same or subclass of type, add it to the result
+                if (resourceType != null && (resourceType == type || resourceType.IsSubclassOf(type)))
+                {
+                    resultGuids.Add(guid);
+                    resultPaths.Add(GetPathFromGUID(guid));
+                }
+            }
+
+            CacheHit cacheHit = new CacheHit();
+            cacheHit.pathCache = resultPaths.ToArray();
+            cacheHit.guidsCache = resultGuids.ToArray();
+
+            //Save in the cache
+            _typeResultCache.Add(type, cacheHit);
+
+            return cacheHit;
+        }
+        #endregion
+
         /// <summary>
         /// Returns true if the AddressablesPathDictionary holds no data whatsoever.
         /// </summary>
@@ -94,106 +386,20 @@ namespace RoR2.Editor
         }
 
         /// <summary>
-        /// Returns an array of all the Addressables Paths stored in the dictionary.
+        /// Returns a ReadOnlyCollection of all the paths stored in the dictionary
         /// </summary>
-        public ReadOnlyArray<string> GetAllPaths()
+        public ReadOnlyCollection<string> GetAllPaths()
         {
-            return new ReadOnlyArray<string>(paths);
+            return new ReadOnlyCollection<string>(paths);
         }
 
         /// <summary>
-        /// Returns an array of all the Addressables Paths stored in the dictionary, from which their assets inherit from <paramref name="t"/>
-        /// <br></br>
-        /// For example, GetAllPathsOfType(typeof(ItemDef)) will return all paths that represent ItemDefs.
+        /// Returns a ReadOnlyCollection of all the guids stored in the dictionary
         /// </summary>
-        /// <param name="t">The type of asset</param>
-        public ReadOnlyArray<string> GetAllPathsOfType(Type t)
+        /// <returns></returns>
+        public ReadOnlyCollection<string> GetAllGUIDS()
         {
-            return GetAllPathsOfTypes(new Type[] { t });
-        }
-
-        /// <summary>
-        /// Returns an array of all the Addressables Paths stored in the dictionary, from which their assets inherit from any of the types specified in <paramref name="types"/>
-        /// </summary>
-        /// <br></br>
-        /// For example, GetAllGUIDSOfType(new Type[] { typeof(ItemDef), typeof(EquipmentDef) }) will return all paths that represent either ItemDefs OR EquipmentDefs.
-        /// <param name="types">The types of assets.</param>
-        public ReadOnlyArray<string> GetAllPathsOfTypes(Type[] types)
-        {
-            if (types == null || types.Length == 0)
-                return GetAllPaths();
-
-            List<string> entries = new();
-            foreach (var key in paths)
-            {
-                var resourceLocations = Addressables.LoadResourceLocationsAsync(key).WaitForCompletion();
-                var resourceLocation = resourceLocations.FirstOrDefault();
-                var resourceType = resourceLocation?.ResourceType;
-
-                if (resourceType != null && (types.Contains(resourceType) || types.Any(t => resourceType.IsSubclassOf(t))))
-                {
-                    entries.Add(key);
-                }
-            }
-            return new ReadOnlyArray<string>(entries.ToArray());
-        }
-
-        /// <summary>
-        /// Returns an array of all the Addressables GUIDS stored in the dictionary.
-        /// </summary>
-        public ReadOnlyArray<string> GetAllGUIDS()
-        {
-            return new ReadOnlyArray<string>(guids);
-        }
-
-        /// <summary>
-        /// Returns an array of all the Addressables GUIDS stored in the dictionary, from which their assets inherit from <paramref name="t"/>
-        /// <br></br>
-        /// For example, GetAllGUIDSOfType(typeof(ItemDef)) will return all GUIDS that represent ItemDefs.
-        /// </summary>
-        /// <param name="t">The type of asset</param>
-        public ReadOnlyArray<string> GetAllGUIDSOfType(Type t)
-        {
-            if (t == null)
-                return GetAllGUIDS();
-
-            List<string> entries = new();
-            foreach (var key in guids)
-            {
-                var resourceLocation = Addressables.LoadResourceLocationsAsync(key).WaitForCompletion().FirstOrDefault();
-                var resourceType = resourceLocation?.ResourceType;
-
-                if (resourceType != null && (resourceType == t || resourceType.IsSubclassOf(t)))
-                {
-                    entries.Add(key);
-                }
-            }
-            return new ReadOnlyArray<string>(entries.ToArray());
-        }
-
-        /// <summary>
-        /// Returns an array of all the Addressables GUIDS stored in the dictionary, from which their assets inherit from any of the types specified in <paramref name="types"/>
-        /// </summary>
-        /// <br></br>
-        /// For example, GetAllGUIDSOfType(new Type[] { typeof(ItemDef), typeof(EquipmentDef) }) will return all GUIDS that represent either ItemDefs OR EquipmentDefs.
-        /// <param name="types">The types of assets.</param>
-        public ReadOnlyArray<string> GetAllGUIDSOfTypes(Type[] types)
-        {
-            if (types == null || types.Length == 0)
-                return GetAllGUIDS();
-
-            List<string> entries = new();
-            foreach (var key in guids)
-            {
-                var resourceLocation = Addressables.LoadResourceLocationsAsync(key).WaitForCompletion().FirstOrDefault();
-                var resourceType = resourceLocation?.ResourceType;
-
-                if (resourceType != null && (types.Contains(resourceType) || types.Any(t => resourceType.IsSubclassOf(t))))
-                {
-                    entries.Add(key);
-                }
-            }
-            return new ReadOnlyArray<string>(entries.ToArray());
+            return new ReadOnlyCollection<string>(guids);
         }
 
         /// <summary>
@@ -237,14 +443,51 @@ namespace RoR2.Editor
         {
             return guidToPathDictionary[guid];
         }
+        #region Obsolete
+        [Obsolete("Create a new instance of \"EntryLookup\" and call \"PerformLookup\" instead")]
+        public ReadOnlyArray<string> GetAllPathsOfType(Type t)
+        {
+            return GetAllPathsOfTypes(new Type[] { t });
+        }
 
+
+        [Obsolete("Create a new instance of \"EntryLookup\" and call \"PerformLookup\" instead")]
+        public ReadOnlyArray<string> GetAllPathsOfTypes(Type[] types)
+        {
+            var lookup = new EntryLookup()
+                .WithLookupType(EntryType.Path)
+                .WithTypeRestriction(types);
+            return new ReadOnlyArray<string>(lookup.PerformLookup().ToArray());
+        }
+
+
+        [Obsolete("Create a new instance of \"EntryLookup\" and call \"PerformLookup\" instead")]
+        public ReadOnlyArray<string> GetAllGUIDSOfType(Type t)
+        {
+            return GetAllGUIDSOfTypes(new Type[] { t });
+        }
+
+
+        [Obsolete("Create a new instance of \"EntryLookup\" and call \"PerformLookup\" instead")]
+        public ReadOnlyArray<string> GetAllGUIDSOfTypes(Type[] types)
+        {
+            var lookup = new EntryLookup()
+                .WithLookupType(EntryType.Guid)
+                .WithTypeRestriction(types);
+            return new ReadOnlyArray<string>(lookup.PerformLookup().ToArray());
+        }
+        #endregion
+
+        private AddressablesPathDictionary() { }
         private AddressablesPathDictionary(Dictionary<string, string> pathToGUID, Dictionary<string, string> guidToPath)
         {
             pathToGUIDDictionary = pathToGUID;
             paths = pathToGUID.Keys.ToArray();
 
             guidToPathDictionary = guidToPath;
-            guids = guidToPath.Keys.ToArray();;
+            guids = guidToPath.Keys.ToArray();
+
+            _typeResultCache = new Dictionary<Type, CacheHit>();
         }
     }
 }